Refactor: split task checkbox and title component
