+++ src/lib/components/atoms/TaskTitle.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+	export let title: string;
+	export let id: number;
+</script>
+
+<a href={`./task/${id}`}>
+	{title}
+</a>
